requirements.txt
Django==1.7.2
Markdown==2.4.1
Pygments==1.6
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.17
django-filter==0.7
django-model-utils==2.0.3
djangorestframework==2.3.14
flower==0.7.0
ipython==2.1.0
kombu==3.0.19
pytz==2014.4
redis==2.10.1
requests==2.3.0
sh==1.09
tornado==3.2.2
uWSGI==2.0.6
wsgiref==0.1.2
psycopg2==2.5.3
newrelic==2.24.0.21
Pillow==2.5.3
subprocess32==3.2.6
jmespath==0.4.1
path.py==5.2
boto==2.35.1
botocore==0.62.0
awscli==1.4.2
hashids==1.0.1
django-debug-panel==0.8.0
django-debug-toolbar==1.2.2
celery==3.1.17
django-celery==3.1.16
flake8==2.3.0
pyflakes==0.8.1
seacucumber==1.5.1
dkimpy==0.5.4
html2text==2014.12.29
ago==0.0.6
